DatePicker表示処理を共通化し、書籍追加画面にも実装
