+++ app/src/main/java/com/android/bookmanager_kotlin/activity/AddBookActivity.kt
@@ -1,10 +1,11 @@
 package com.android.bookmanager_kotlin.activity
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import com.android.bookmanager_kotlin.R
+import com.android.bookmanager_kotlin.util.DatePickerUtils.showDatePicker
+import kotlinx.android.synthetic.main.activity_add_book.*
 
 class AddBookActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -12,6 +13,10 @@ class AddBookActivity : AppCompatActivity() {
         setContentView(R.layout.activity_add_book)
 
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
+
+        et_add_book_purchase_date.setOnClickListener {
+            showDatePicker(this, et_add_book_purchase_date)
+        }
     }
 
     // アクションバー戻るボタンクリック処理
