+++ app/src/main/java/com/android/bookmanager_kotlin/fragment/EditBookFragment.kt
@@ -1,20 +1,17 @@
 package com.android.bookmanager_kotlin.fragment
 
-import android.app.DatePickerDialog
-import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.setFragmentResultListener
 import com.android.bookmanager_kotlin.R
+import com.android.bookmanager_kotlin.util.DatePickerUtils.showDatePicker
 import com.android.bookmanager_kotlin.util.FragmentUtils
 import kotlinx.android.synthetic.main.fragment_edit_book.*
-import java.util.*
 
 class EditBookFragment : Fragment() {
 
@@ -54,18 +51,5 @@ class EditBookFragment : Fragment() {
             FragmentUtils().showFragment(BookListFragment(), parentFragmentManager, R.id.fl_activity_book_list)
         }
         return super.onOptionsItemSelected(item)
-    }
-
-    private fun showDatePicker(context: Context, editText: EditText) {
-        val calendar = Calendar.getInstance()
-        val year = calendar.get(Calendar.YEAR)
-        val month = calendar.get(Calendar.MONTH)
-        val dayOfMonth = calendar.get(Calendar.DAY_OF_MONTH)
-
-        DatePickerDialog(
-            context, { _, y, m, d ->
-                editText.setText("$y/${m + 1}/$d")
-            }, year, month, dayOfMonth
-        ).show()
     }
 }
